Merged in fix-backend-allow-adding-env-without-default (pull request #323)
* Allow to add `env` w/o default defined

This is the original fix that was introduced by Aleksander Ambrożkiewicz
in the IMAX project's clone of the old `aws-boilerplate`.
However, until now, it was not propagated into neither `aws-boilerplate`
nor `saas-boilerplate`.

This issue prevented the developer to add an environment variable to
`settings.py` w/o the `default` kwarg, as that caused the docker build
to fail, even if that specific environment variables was defined in the
local `.env` (!) or `.test.env`.

It's not the perfect fix, because the expected behavior should be to
just use the `env_file` property defined in the
`docker-compose.override.yml` instead of the requirement to provide the
separate placeholder environment variables' definition before that step.

However, this is the quick workaround that for now proved to be working.

Approved-by: Krzysztof Stronka