Merged in upgrade-workers-dependencies (pull request #331)
Update `services/workers` dependencies

* Use `pdm sync` instead of `pdm install`

The latter updates the lock, which is not something desirable during the
build & deploy.

* Adjust Python version specifier

Using `>=3.8` causes the PDM to use any newer MINOR version that is
available on the system, which not always can be desired. For example,
on my localhost, PDM tries to use Python `3.10`, while the
saas-boilerplate is at the moment tested and expected to run on `3.8.*`.

As such, it's recommended to disallow PDM from using any version higher
than `3.8.*` until we actually migrate to using a more recent MINOR
version
for saas-boilerplate.

* Update SQLAlchemy MINOR version

This change includes:
- Handle backward-incompatible bugfix in `Dialect.has_table()`
    <https://docs.sqlalchemy.org/en/14/changelog/changelog_14.html#change-80d8fab2614090c6f7e1304727917693>
- Fix deprecation warning
    ```
    SADeprecationWarning: Calling URL() directly is deprecated and will
    be disabled in a future release.  The public constructor for URL is
    now the URL.create() method.
    ```

* Update packages

- Update moto 3.1.9 -> 3.1.10
- Update boto3 1.23.4 -> 1.23.6
- Update botocore 1.26.4 -> 1.26.6
- Update pytest-factoryboy 2.3.0 -> 2.3.1

* Update `black` and add version specifiers for all deps

Approved-by: Krzysztof Stronka